refactor(modules): extract showValidationError and resetValidationError instance methods from main module
