+++ src/instanceMethods.js
@@ -1,3 +1,4 @@
 export * from './instanceMethods/hideLoading'
 export * from './instanceMethods/getInput'
 export * from './instanceMethods/enable-disable-elements'
+export * from './instanceMethods/show-set-validation-error'
