+++ src/sweetalert2.js
@@ -629,43 +629,6 @@ Object.keys(instanceMethods).forEach(key => {
   }
 })
 
-// Show block with validation error
-sweetAlert.showValidationError = (error) => {
-  if (currentContext) {
-    const {domCache} = currentContext
-    domCache.validationError.innerHTML = error
-    const popupComputedStyle = window.getComputedStyle(domCache.popup)
-    domCache.validationError.style.marginLeft = `-${popupComputedStyle.getPropertyValue('padding-left')}`
-    domCache.validationError.style.marginRight = `-${popupComputedStyle.getPropertyValue('padding-right')}`
-    dom.show(domCache.validationError)
-
-    const input = sweetAlert.getInput()
-    if (input) {
-      input.setAttribute('aria-invalid', true)
-      input.setAttribute('aria-describedBy', swalClasses.validationerror)
-      dom.focusInput(input)
-      dom.addClass(input, swalClasses.inputerror)
-    }
-  }
-}
-
-// Hide block with validation error
-sweetAlert.resetValidationError = () => {
-  if (currentContext) {
-    const {domCache} = currentContext
-    if (domCache.validationError) {
-      dom.hide(domCache.validationError)
-    }
-
-    const input = sweetAlert.getInput()
-    if (input) {
-      input.removeAttribute('aria-invalid')
-      input.removeAttribute('aria-describedBy')
-      dom.removeClass(input, swalClasses.inputerror)
-    }
-  }
-}
-
 sweetAlert.getProgressSteps = () => {
   if (currentContext) {
     const {params} = currentContext
